pyproject.toml
[tool.poetry]
name = "streamlit-proj-1"
version = "0.1.0"
description = ""
authors = ["fspjonny <fabiosilvapedro@yahoo.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit-pydantic = "^0.6.0"
pydantic-settings = "^2.6.1"
groq = "^0.13.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
